Add status icons and other updates
Add status icons provided by AGL graphics design team, and tweak
the animation behavior to simulate power on at the start of the
cycle by blinking them.  Additionally, the speed has been switched
to MPH, and the turn signal animation has been udpated to simulate
passing as opposed to a turn at the start of the cycle.

Change-Id: If5e3b9d2138d1129a7478bea1f4f2cb6b05f1499
Signed-off-by: Scott Murray <scott.murray@konsulko.com>