Update VIS vehicle signal support
Rework to change the VIS signals being used for the cruise control
and display mode switching to new AGL custom steering wheel switch
signals.  In practice mapping multiple CAN signals from the LIN
polling to a couple VIS cruise control actuator signals does not
work well with the behavior of the CAN feeder from KUKSA.val.
With the separate steering wheel switch signals and duplicate
filtering in libqtappfw, things look more like they did with
signal-composer and should hopefully be robust.  This should be
revisited if a custom LIN signal feeder becomes an option, as
overall it would be better to leverage as much of the standard VSS
schema as possible.

Bug-AGL: SPEC-4409

Change-Id: I6369b030e282a8040b144d6384daeec46dc7098d
Signed-off-by: Scott Murray <scott.murray@konsulko.com>